* added package for globals * `DataBases`, `StorageEngine` are singletons now * added `private Kryo kryo` field to `RowIterator` * fixed a little logic * added `apache commons-io` for working with files(maybe it will be not used)
