+++ src/main/java/com/dbms/gui/MainController.java
@@ -1,6 +1,5 @@
 package com.dbms.gui;
 
-import com.dbms.logic.ExpressionExecutor;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.TextArea;
