+++ src/main/java/com/dbms/storage/StorageEngine.java
@@ -1,10 +1,8 @@
 package com.dbms.storage;
 
-import com.dbms.data_types.Char;
-import com.dbms.data_types.Varchar;
+import com.dbms.global.Global;
 import com.dbms.structs.DataBase;
 import com.dbms.structs.Table;
-import com.dbms.structs.TypeDescription;
 import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.io.Input;
 import com.esotericsoftware.kryo.io.Output;
@@ -17,54 +15,51 @@ import java.nio.file.Paths;
 import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
 import java.util.Map;
 
 public class StorageEngine {
 
-    private static String dataBaseRootDirectory = "db";
-    private static String tableHeaders = getFilePath("meta_data", "table_headers");
-    private static String tableData = getFilePath("data", "table_data");
+    private StorageEngine() {
+    }
 
-    private static String[] dbDirectories = {//full paths of service folders
+    private static StorageEngine instance;
+
+    public static StorageEngine getInstance() {
+        if (instance == null)
+            instance = new StorageEngine();
+
+        return instance;
+    }
+
+    private String dataBaseRootDirectory = "db";
+    private String tableHeaders = getFilePath("meta_data", "table_headers");
+    private String tableData = getFilePath("data", "table_data");
+
+    private String[] dbDirectories = {//full paths of service folders
             tableHeaders,
             tableData
     };
 
-    private static String tmpFileExtension = ".tmp";
+    private String tmpFileExtension = ".tmp";
 
-    private StorageEngine() {
-    }
+    private Kryo kryo = Global.getKryo();
 
-    public static Kryo kryo = new Kryo();
-
-    public static void init() throws Exception {
+    public void init() throws Exception {
         var dbrd = new File(dataBaseRootDirectory);
         if (!dbrd.isDirectory())
             if (!dbrd.mkdir())
                 throw new Exception("Cannot create database root directory");
-
-        kryo.register(ArrayList.class);
-        kryo.register(LinkedHashMap.class);
-        kryo.register(HashMap.class);
-
-        kryo.register(Char.class);
-        kryo.register(Varchar.class);
-
-        kryo.register(TypeDescription.class);
-
-        kryo.register(Table.class);
     }
 
-    private static String getFilePath(String first, String... more) {
+    private String getFilePath(String first, String... more) {
         return Paths.get(first, more).toString();
     }
 
-    private static String db_path(String name) {
+    private String db_path(String name) {
         return getFilePath(dataBaseRootDirectory, name);
     }
 
-    public static DataBase readDataBase(File dataBaseDirectory) throws Exception {
+    public DataBase readDataBase(File dataBaseDirectory) throws Exception {
         var dataBaseName = dataBaseDirectory.getName();
         if (!dataBaseDirectory.isDirectory())
             throw new Exception(String.format("The '%s' database does not exist", dataBaseName));
@@ -77,11 +72,11 @@ public class StorageEngine {
         return new DataBase(dataBaseName, tables);
     }
 
-    public static DataBase readDataBase(String name) throws Exception {
+    public DataBase readDataBase(String name) throws Exception {
         return readDataBase(new File(db_path(name)));
     }
 
-    public static Map<String, DataBase> readAllDataBases() throws Exception {
+    public Map<String, DataBase> readAllDataBases() throws Exception {
         var dataBases = new HashMap<String, DataBase>();
 
         for (var f : new File(dataBaseRootDirectory).listFiles(File::isDirectory))
@@ -90,7 +85,7 @@ public class StorageEngine {
         return dataBases;
     }
 
-    public static void initDataBase(String name) throws Exception {
+    public void initDataBase(String name) throws Exception {
         var db_path = db_path(name);
 
         if (new File(db_path).exists())
@@ -113,40 +108,40 @@ public class StorageEngine {
         }
     }
 
-    public static void writeTableHeader(String dataBaseName, Table table) throws Exception {
+    public void writeTableHeader(String dataBaseName, Table table) throws Exception {
         try (var out = new Output(new FileOutputStream(getFilePath(db_path(dataBaseName), tableHeaders, table.getTableName())))) {
             kryo.writeObject(out, table);
         }
     }
 
-    public static void initTableData(String dataBaseName, String tableName) throws Exception {
+    public void initTableData(String dataBaseName, String tableName) throws Exception {
         try (var ignored = new FileOutputStream(getFilePath(db_path(dataBaseName), tableData, tableName))) {
         }
     }
 
-    public static Table readTable(File file) throws Exception {
+    public Table readTable(File file) throws Exception {
         try (var in = new Input(new FileInputStream(file))) {
             return kryo.readObject(in, Table.class);
         }
     }
 
-    public static void appendRow(String dataBaseName, String tableName, ArrayList row) throws Exception {
+    public void appendRow(String dataBaseName, String tableName, ArrayList row) throws Exception {
         try (var out = new Output(new FileOutputStream(getFilePath(db_path(dataBaseName), tableData, tableName), true))) {
             kryo.writeObject(out, row);
         }
     }
 
-    public static RowIterator getRowIterator(String dataBaseName, String tableName) throws Exception {
+    public RowIterator getRowIterator(String dataBaseName, String tableName) throws Exception {
         return new RowIterator(new FileInputStream(getFilePath(db_path(dataBaseName), tableData, tableName)));
     }
 
-    public static String createTempTableData(String dataBaseName, String tableName) throws Exception {
+    public String createTempTableData(String dataBaseName, String tableName) throws Exception {
         var tempTableName = tableName + tmpFileExtension;
         initTableData(dataBaseName, tempTableName);
         return tempTableName;
     }
 
-    public static void replaceTableDataToTemp(String dataBaseName, String tableName) throws Exception {
+    public void replaceTableDataToTemp(String dataBaseName, String tableName) throws Exception {
         var p = Paths.get(getFilePath(db_path(dataBaseName), tableData, tableName + tmpFileExtension));
         Files.move(p, p.resolveSibling(tableName), StandardCopyOption.REPLACE_EXISTING);
     }
